antigravity_core/requirements.txt
google-genai>=0.2.0
ffmpeg-python
openai-whisper
git+https://github.com/m-bain/whisperx.git
pydantic
python-dotenv
rich
watchdog
